nsm.h: change getpid to geppid
Adljack starts itself in a terminal application. The terminal application obtains
the same pid as is started by the NSM daemon. This is the pid the NSM Daemon 'owns'.

To announce with the same pid as the process started by the NSM daemon, it's needed
to use getppid (get parent pid), this is the pid of the terminal, whereas
getpid gets us the pid of adljack.