common.cc: add SIGHUP to signals
When under NSM session management, a NSM clients needs to exite when it gets a SIGTERM
signal from the NSM daemon. The case of adljack is a special one, as it launches
itself in a terminal application.

To match the pid process of the NSM daemon when it starts adljack, the pid of the
terminal application is send (parent pid). So the terminal applications get the
SIGTERM signal from the NSM daemon.

SIGHUP is a signal sent to a process when its controlled terminal is closed.
So by adding SIGHUP to the catched signals, adljack exites when the
terminal application adljack is launched in, gets a SIGTERM signal.